+++ packages/providers/lib.esm/url-json-rpc-provider.js
@@ -1,4 +1,13 @@
 "use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
 import { defineReadOnly, getStatic } from "@ethersproject/properties";
 import { Logger } from "@ethersproject/logger";
 import { version } from "./_version";
@@ -20,6 +29,11 @@ export class UrlJsonRpcProvider extends JsonRpcProvider {
                 defineReadOnly(this, key, apiKey[key]);
             });
         }
+    }
+    detectNetwork() {
+        return __awaiter(this, void 0, void 0, function* () {
+            return this.network;
+        });
     }
     _startPending() {
         logger.warn("WARNING: API provider does not support pending filters");